requirements/base.txt
Django>=5.2
psycopg2-binary
SQLAlchemy>=2.0
alembic
python-dotenv
pydantic
python-dateutil
djangorestframework
djangorestframework-simplejwt
requests


